py/requirements.txt
protobuf==3.6.0
grpcio-tools
googleapis-common-protos
numpy